Add new files

+++ client/src/pages/Home/Home.jsx
@@ -1,6 +1,7 @@
+/* eslint-disable react-hooks/exhaustive-deps */
 /* eslint-disable no-unused-vars */
 import { useState, useEffect } from "react";
-import Navebar from "../../components/Navbar/Navbar";
+import Navbar from "../../components/Navbar/Navbar"; 
 import NoteCard from "../../components/Cards/NoteCard";
 import { MdAdd } from "react-icons/md";
 import AddEditNotes from "../../components/AddEditNotes/AddEditNotes";
@@ -11,6 +12,7 @@ import Toast from "../../components/ToastMessage/Toast";
 import EmptyCard from "../../components/Cards/EmptyCard";
 import AddNoteImg from "../../assets/images/AddNoteImg.svg";
 import NoDataImg from "../../assets/images/no-data-icon.webp";
+
 const Home = () => {
   const [openAddEditModal, setOpenAddEditModal] = useState({
     isShown: false,
@@ -21,24 +23,24 @@ const Home = () => {
   const [showToastMsg, setShowToastMsg] = useState({
     isShown: false,
     type: "add",
-    data: null,
+    message: "",
   });
-
-  useEffect(() => {
-    Modal.setAppElement("#root");
-  }, []);
 
   const [allNotes, setAllNotes] = useState([]);
   const [userInfo, setUserInfo] = useState(null);
   const [isSearch, setIsSearch] = useState(false);
   const navigate = useNavigate();
 
-  //handle edit
+  useEffect(() => {
+    Modal.setAppElement("#root");
+    getAllNotes();
+    getUserInfo();
+  }, []);
+
   const handleEdit = (noteDetails) => {
     setOpenAddEditModal({ isShown: true, data: noteDetails, type: "edit" });
   };
 
-  //Handle toast Message
   const showToastMessage = (message, type) => {
     setShowToastMsg({
       isShown: true,
@@ -47,7 +49,6 @@ const Home = () => {
     });
   };
 
-  //handle toast close
   const handleCloseToast = () => {
     setShowToastMsg({
       isShown: false,
@@ -55,7 +56,6 @@ const Home = () => {
     });
   };
 
-  //Get use Info
   const getUserInfo = async () => {
     try {
       const response = await axiosInstance.get("/get-user");
@@ -63,55 +63,42 @@ const Home = () => {
         setUserInfo(response.data.user);
       }
     } catch (error) {
-      if (error.response.status === 401) {
+      if (error.response && error.response.status === 401) {
         localStorage.clear();
         navigate("/login");
       }
     }
   };
 
-  //Get All Notes
   const getAllNotes = async () => {
     try {
       const response = await axiosInstance.get("/get-all-notes");
-
       if (response.data && response.data.notes) {
         setAllNotes(response.data.notes);
       }
     } catch (error) {
-      console.log("data not found");
+      console.log("Data not found");
     }
   };
 
-  //Delete Note
   const deleteNote = async (data) => {
     const noteId = data._id;
     try {
       const response = await axiosInstance.delete("/delete-note/" + noteId);
-
       if (response.data && !response.data.error) {
         showToastMessage("Note Deleted Successfully", "delete");
         getAllNotes();
       }
     } catch (error) {
-      //Handle login error
-      if (
-        error.response &&
-        error.response.data &&
-        error.response.data.message
-      ) {
-        console.log("Unexpected Error");
-      }
+      console.log("Unexpected Error");
     }
   };
 
-  //Search for a note
   const onSearchNote = async (query) => {
     try {
       const response = await axiosInstance.get("/search-notes", {
         params: { query },
       });
-
       if (response.data && response.data.notes) {
         setIsSearch(true);
         setAllNotes(response.data.notes);
@@ -121,7 +108,6 @@ const Home = () => {
     }
   };
 
-  //Update Pinned
   const updateIsPinned = async (noteData) => {
     const noteId = noteData._id;
     try {
@@ -131,9 +117,8 @@ const Home = () => {
           isPinned: !noteData.isPinned,
         }
       );
-
       if (response.data && response.data.note) {
-        showToastMessage("isPinned Updated Successfully");
+        showToastMessage("isPinned Updated Successfully", "update");
         getAllNotes();
       }
     } catch (error) {
@@ -146,24 +131,17 @@ const Home = () => {
     getAllNotes();
   };
 
-  useEffect(() => {
-    getAllNotes();
-    getUserInfo();
-
-    return () => {};
-  }, []);
-
   return (
     <>
-      <Navebar
+      <Navbar
         userInfo={userInfo}
         onSearchNote={onSearchNote}
         handleClearSearch={handleClearSearch}
       />
 
-      <div className="container w-[90%] ml-20">
+      <div className="container mx-auto px-4 sm:px-6 lg:px-8">
         {allNotes.length > 0 ? (
-          <div className="grid grid-cols-3 gap-4 mt-8">
+          <div className="grid grid-cols-2 sm:grid-cols-3 lg:grid-cols-4 gap-4 mt-8">
             {allNotes.map((item) => (
               <NoteCard
                 key={item._id}
@@ -183,20 +161,20 @@ const Home = () => {
             imgSrc={isSearch ? NoDataImg : AddNoteImg}
             message={
               isSearch
-                ? `Oops! no notes found matching your search`
-                : `Start creating your firest note! Click the 'Add' button to jot down your thought, ideas, and, remiders. Lest's get started:`
+                ? `Oops! No notes found matching your search`
+                : `Start creating your first note! Click the 'Add' button to jot down your thoughts, ideas, and reminders. Let's get started:`
             }
           />
         )}
       </div>
 
       <button
-        className="w-16 h-16 flex items-center justify-center rounded-2xl bg-primary hover:bg-blue-600 absolute right-10 bottom-10"
+        className="w-14 h-14 sm:w-16 sm:h-16 flex items-center justify-center rounded-full bg-primary hover:bg-blue-600 fixed right-6 bottom-6 sm:right-10 sm:bottom-10"
         onClick={() => {
           setOpenAddEditModal({ isShown: true, type: "add", data: null });
         }}
       >
-        <MdAdd className="text-[32px] text-white" />
+        <MdAdd className="text-[24px] sm:text-[32px] text-white" />
       </button>
 
       <Modal
@@ -210,7 +188,7 @@ const Home = () => {
           },
         }}
         contentLabel=""
-        className="w-[40%] max-h-3/4 bg-white rounded-md mx-auto mt-14 p-5 overflow-scroll"
+        className="w-full sm:w-[60%] lg:w-[40%] max-h-3/4 bg-white rounded-md mx-auto mt-14 p-5 overflow-y-auto"
       >
         <AddEditNotes
           type={openAddEditModal.type}
